+++ test/hexapdf/test_document.rb
@@ -436,17 +436,17 @@ describe HexaPDF::Document do
     end
 
     it "validates indirect objects" do
-      @doc.add(Type: :Catalog)
+      obj = @doc.add(Type: :Catalog)
       refute(@doc.validate(auto_correct: false))
 
       called = false
-      assert(@doc.validate { called = true })
+      assert(@doc.validate {|o| assert_same(obj, o); called = true })
       assert(called)
     end
 
     it "validates the trailer object" do
       @doc.trailer[:ID] = :Symbol
-      refute(@doc.validate)
+      refute(@doc.validate {|obj| assert_same(@doc.trailer, obj) })
     end
   end